Fixes evergreen/rails to work out of the box in an engine.
In an engine, the normal behaviour is to place your tests in
spec/javascripts or test/javascripts.

However, when the app is booted, the dummy app in spec/dummy or
test/dummy is booted, so Rails.root returns <engine_path>/spec/dummy.

In this case, we want evergreen to look at <engine_path>, so we go up
two levels from the Rails.root.

I've also added some tests for the rails.rb file, as there were none
before, and tests for this new functionality. Unfortunately it requires
rails to be a development dependency since we make use of Rails::Engine
and Rails::Application in the tests.

I couldn't find a better way to check if we're in an engine's dummy app
than to look at the #engine_name - this is not great but there is no
other way I could find.

Also, I couldn't get the entire test suite to pass after a fresh
checkout, so I'm not sure if this affects anything else.