Fix to make sure user specific configuration is loaded before the application is built.
We were having an issue where we were overriding public_dir in a
.evergreen file. When the server is booted, the application was being built
with the default settings.

When a new suite instance was created, the user config was being loaded,
but since the Evergreen.application instance was cached, its public_dir
setting was not getting updated.

I've move the user config loading to a load_user_config! method on
Evergreen, and called this before the application is built to make sure
the user config gets loaded first.