Introduce noConflict for global JS methods
Attempting to support a noConflict option similar to jQuery's
noConflict. If Evergreen nukes a pre-existing method like `require`,
calling `Evergreen.noConflict()` will retrive the old definition, and
the developer is required to use `Evergreen.require` instead of `require`

The code is pretty crude, but any sort of refactor seems to upset
poltergeist, so I'll get to refactoring this later

This should satisfy issue #17.